+++ inferno-core/src/Inferno/Core.hs
@@ -9,30 +9,31 @@ import Control.Monad (foldM)
 import Control.Monad.Catch (MonadCatch, MonadThrow)
 import Control.Monad.Except (MonadFix)
 import Data.Bifunctor (bimap)
-import qualified Data.List.NonEmpty as NEList
+import Data.List.NonEmpty (NonEmpty)
 import qualified Data.Map as Map
 import qualified Data.Set as Set
 import Data.Text (Text)
 import Inferno.Eval (TermEnv, eval, runEvalM)
 import Inferno.Eval.Error (EvalError)
-import Inferno.Infer (inferExpr, inferTypeReps)
+import Inferno.Infer (TypeError, inferExpr, inferTypeReps)
+import Inferno.Infer.Error (Location)
 import Inferno.Infer.Pinned (pinExpr)
 import Inferno.Module (Module (..))
 import Inferno.Module.Builtin (builtinModule)
-import Inferno.Module.Prelude (ModuleMap, baseOpsTable, builtinModulesOpsTable, builtinModulesPinMap, builtinModulesTerms)
-import Inferno.Parse (parseExpr, prettyError)
-import Inferno.Types.Syntax (Expr (App, TypeRep), ExtIdent, SourcePos, collectArrs)
+import Inferno.Module.Prelude (ModuleMap, baseOpsTable, builtinModulesOpsTable, builtinModulesPinMap, builtinModulesTerms, preludeNameToTypeMap)
+import Inferno.Parse (InfernoParsingError, parseExpr)
+import Inferno.Types.Syntax (Comment, CustomType, Expr (App, TypeRep), ExtIdent, ModuleName, Namespace, SourcePos, TypeClass, TypeMetadata, collectArrs)
 import Inferno.Types.Type (ImplType (ImplType), TCScheme (ForallTC))
 import Inferno.Types.Value (ImplEnvM, Value, runImplEnvM)
 import Inferno.Types.VersionControl (Pinned, VCObjectHash, pinnedToMaybe)
 import Inferno.VersionControl.Types (VCObject (VCFunction))
 import Prettyprinter (Pretty)
-import Text.Megaparsec (initialPos)
+import Text.Megaparsec (ParseError, initialPos)
 
 data InfernoError
-  = ParseError String
-  | PinError String
-  | InferenceError String
+  = ParseError (NonEmpty (ParseError Text InfernoParsingError, SourcePos))
+  | PinError [TypeError SourcePos]
+  | InferenceError [TypeError SourcePos]
   deriving (Eq, Show)
 
 -- | Public API for the Inferno interpreter.
@@ -53,7 +54,7 @@ data Interpreter m c = Interpreter
       Either InfernoError (Expr (Maybe VCObjectHash) SourcePos),
     parseAndInfer ::
       Text ->
-      Either InfernoError (Expr (Pinned VCObjectHash) SourcePos, TCScheme),
+      Either InfernoError (Expr (Pinned VCObjectHash) SourcePos, TCScheme, Map.Map (Location SourcePos) (TypeMetadata TCScheme), [Comment SourcePos]),
     -- | Evaluates all functions in given closure and creates a pinned env containing them
     mkEnvFromClosure ::
       Map.Map ExtIdent (Value c (ImplEnvM m c)) ->
@@ -61,11 +62,17 @@ data Interpreter m c = Interpreter
       ImplEnvM m c (TermEnv VCObjectHash c (ImplEnvM m c)),
     -- | The default pinned env containing only the prelude
     defaultEnv ::
-      TermEnv VCObjectHash c (ImplEnvM m c)
+      TermEnv VCObjectHash c (ImplEnvM m c),
+    -- | The type of each name in this interpreter's prelude
+    nameToTypeMap ::
+      Map.Map (Maybe ModuleName, Namespace) (TypeMetadata TCScheme),
+    -- | The set of all type classes in this interpreter's prelude
+    typeClasses ::
+      Set.Set TypeClass
   }
 
-mkInferno :: forall m c. (MonadThrow m, MonadCatch m, MonadFix m, Eq c, Pretty c) => ModuleMap m c -> m (Interpreter m c)
-mkInferno prelude = do
+mkInferno :: forall m c. (MonadThrow m, MonadCatch m, MonadFix m, Eq c, Pretty c) => ModuleMap m c -> [CustomType] -> m (Interpreter m c)
+mkInferno prelude customTypes = do
   -- We pre-compute envs that only depend on the prelude so that they can be
   -- shared among evaluations of different scripts
   (preludeIdentEnv, preludePinnedEnv) <- runImplEnvM Map.empty $ builtinModulesTerms prelude
@@ -75,34 +82,36 @@ mkInferno prelude = do
         parseAndInferTypeReps = parseAndInferTypeReps,
         parseAndInfer = parseAndInfer,
         mkEnvFromClosure = mkEnvFromClosure preludePinnedEnv,
-        defaultEnv = (preludeIdentEnv, preludePinnedEnv)
+        defaultEnv = (preludeIdentEnv, preludePinnedEnv),
+        nameToTypeMap = preludeNameToTypeMap prelude,
+        typeClasses = typeClasses
       }
   where
     parseAndInfer src =
       -- parse
-      case parseExpr (baseOpsTable prelude) (builtinModulesOpsTable prelude) src of
+      case parseExpr (baseOpsTable prelude) (builtinModulesOpsTable prelude) customTypes src of
         Left err ->
-          Left $ ParseError $ (prettyError $ fst $ NEList.head err)
-        Right (ast, _comments) ->
+          Left $ ParseError err
+        Right (ast, comments) ->
           -- pin free variables to builtin prelude function hashes
           case pinExpr (builtinModulesPinMap prelude) ast of
-            Left err -> Left $ PinError $ show err
+            Left err -> Left $ PinError err
             Right pinnedAST ->
               -- typecheck
               case inferExpr prelude pinnedAST of
-                Left err -> Left $ InferenceError $ show err
-                Right (pinnedAST', sch, _tyMap) ->
-                  Right (pinnedAST', sch)
+                Left err -> Left $ InferenceError err
+                Right (pinnedAST', sch, tyMap) ->
+                  Right (pinnedAST', sch, tyMap, comments)
 
     parseAndInferTypeReps src =
       case parseAndInfer src of
         Left err -> Left err
-        Right (pinnedAST', sch@(ForallTC _ _ (ImplType _ typ))) ->
+        Right (pinnedAST', sch@(ForallTC _ _ (ImplType _ typ)), _tyMap, _comments) ->
           let sig = collectArrs typ
            in -- infer runtime type-reps
-              case inferTypeReps allClasses sch (init sig) (last sig) of
+              case inferTypeReps typeClasses sch (init sig) (last sig) of
                 Left err ->
-                  Left $ InferenceError $ show err
+                  Left $ InferenceError err
                 Right runtimeReps ->
                   let finalAst =
                         foldl
@@ -111,7 +120,7 @@ mkInferno prelude = do
                           [TypeRep (initialPos "dummy") ty | ty <- runtimeReps]
                    in Right finalAst
 
-    allClasses = Set.unions $ moduleTypeClasses builtinModule : [cls | Module {moduleTypeClasses = cls} <- Map.elems prelude]
+    typeClasses = Set.unions $ moduleTypeClasses builtinModule : [cls | Module {moduleTypeClasses = cls} <- Map.elems prelude]
 
     -- TODO at some point: instead of evaluating closure and putting into pinned env,
     -- add closure into the expression being evaluated by using let bindings.